+++ src/ethereum/info/sepoliaInfo.ts
@@ -89,7 +89,12 @@ const networkInfo: EthereumNetworkInfo = {
   pluginMnemonicKeyName: 'sepoliaMnemonic',
   pluginRegularKeyName: 'sepoliaKey',
   evmGasStationUrl: null,
-  networkFees
+  networkFees,
+  decoyAddressConfig: {
+    count: 5,
+    minTransactionCount: 10,
+    maxTransactionCount: 100
+  }
 }
 
 const currencyInfo: EdgeCurrencyInfo = {